requirements.txt
# Core dependencies
flask==3.0.0
scapy==2.5.0
requests==2.31.0
pandas==2.1.3
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2

# Network analysis
dnspython==2.4.2
python-whois==0.8.0

# Visualization
plotly==5.18.0
dash==2.14.2

# Logging and monitoring
colorama==0.4.6
python-dotenv==1.0.0

# Async support
aiohttp==3.9.1
